Implemented Drone
The patrol prefab has been modified to now include the drone meshes, with a few custom materials and 2 animations. Its death is supposed to be an explosion with fire particles, however that is yet to be tested as a script must call a SendMessage directly to that drone's Destroyed function.